+++ backend/src/main/java/com/borasio_back/backend/repository/MotoristaRepository.java
@@ -3,5 +3,5 @@ package com.borasio_back.backend.repository;
 import com.borasio_back.backend.model.entity.Motorista;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MotoristaRepository extends JpaRepository<Motorista, Integer> {
+public interface MotoristaRepository extends JpaRepository<Motorista, Long> {
 }
